% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{average_over_instances}
\alias{average_over_instances}
\title{Take the average of a data frame over the instances}
\usage{
average_over_instances(x, instances)
}
\description{
Take the average of a data frame over the instances
}
\author{
Sean Kent
}
\keyword{internal}
