% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_methods.R
\name{build_instance_feature}
\alias{build_instance_feature}
\title{This flatten the MilData type of data to regular multiple instance data where each instance is a vector}
\usage{
build_instance_feature(
  data,
  qtls = seq(0.05, 0.95, length.out = 10),
  mean = TRUE,
  sd = TRUE
)
}
\arguments{
\item{data}{A MilData object.}

\item{qtls}{Quantiles to be extracted from each instance empirical distribution.}

\item{mean}{Whether or not to extract mean.}

\item{sd}{Whether or not to extract median.}
}
\value{
A data.frame that is ready to be used in `MI_SVM()` function.
}
\description{
This flatten the MilData type of data to regular multiple instance data where each instance is a vector by extracting distribution sample quantiles, mean and sd.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
df1 <- build_instance_feature(MilData1, seq(0.05, 0.95, length.out = 10))
}
\author{
Yifei Liu
}
