% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkfm_functions.R
\name{build_kernel_feature_map}
\alias{build_kernel_feature_map}
\title{Build a kernel feature map from fitted object}
\usage{
build_kernel_feature_map(fit, data)
}
\description{
TODO: make documentation
}
\author{
Sean Kent
}
