% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkfm_functions.R
\name{build_kernel_mean_map}
\alias{build_kernel_mean_map}
\title{Build a kernel mean map from fitted object}
\usage{
build_kernel_mean_map(fit, data)
}
\description{
TODO: documentation
}
\author{
Sean Kent
}
