% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_methods.R
\name{build_poly_instance_feature}
\alias{build_poly_instance_feature}
\title{This flattens the MilData type of data to regular multiple instance data where each instance is a vector}
\usage{
build_poly_instance_feature(data, degree = 2, constant = 1)
}
\arguments{
\item{data}{A MilData object.}

\item{degree}{degree of the polynomial kernel, default = 2}

\item{constant}{term added to inner product before applying the degree}
}
\value{
A data.frame that is ready to be used in `MI_SVM()` function.
}
\description{
This flatten the MilData type of data to regular multiple instance data where
each instance is a vector by extracting the features that would be present
in a polynomial kernel with given degree (default = 2).  Each instance_name
has features calculated in the input space based on the underlying kernel
(x'y + c)^d where c = constant and d = degree are input parameters.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
df1 <- build_poly_instance_feature(MilData1, degree = 2)
}
\author{
Sean Kent
}
