% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compute_kernel}
\alias{compute_kernel}
\title{Compute kernel matrix based on type}
\usage{
compute_kernel(x, x2 = NULL, type = "linear", sigma = NULL)
}
\description{
Compute kernel matrix based on type
}
\author{
Sean Kent
}
\keyword{internal}
