% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_y}
\alias{convert_y}
\title{Store the levels of y and convert to 0,1 numeric format.}
\usage{
convert_y(y)
}
\description{
Store the levels of y and convert to 0,1 numeric format.
}
\author{
Sean Kent
}
\keyword{internal}
