% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{cv_MI_SVM}
\alias{cv_MI_SVM}
\title{Cross-validation function for MI_SVM}
\usage{
cv_MI_SVM(
  data,
  n_fold,
  fold_id,
  cost_seq,
  kernel = "radial",
  max.step = 500,
  type = "C-classification"
)
}
\arguments{
\item{data}{Same as in MI_SVM()}

\item{n_fold}{Default to be 5}

\item{fold_id}{Can be skipped if n_fold is set.}

\item{cost_seq}{The cost sequence.}

\item{kernel}{The kernel to be used in MI_SVM()}

\item{max.step}{Maximum steps to be used in MI_SVM()}

\item{type}{type to be used in MI_SVM()}
}
\value{
A list that contains best model, optimal cost value, the AUCs, and the cost sequence.
}
\description{
Cross-validation function for MI_SVM.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
df1 <- build_instance_feature(MilData1, seq(0.05, 0.95, length.out = 10))
foo_cv <- cv_MI_SVM(data = df1, n_fold = 3, cost_seq = 2^(-2:2))
}
\author{
Yifei Liu
}
