% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mil_distribution.R
\name{cv_mild}
\alias{cv_mild}
\title{Function to perform cross validation for the cost used in mil_distribution() function.}
\usage{
cv_mild(
  data,
  n_fold,
  fold_id,
  cost_seq = 2^(-2:2),
  max.step = 300,
  sigma = 0.05,
  ...
)
}
\arguments{
\item{data}{The dataset as a MilData object.}

\item{n_fold}{The number of folds.}

\item{fold_id}{A user specified fold_id. Suggested not to specify
this.}

\item{cost_seq}{The cost sequence, default to be 2^(-2:2).}

\item{max.step}{Same as in mil_distribution(), default to be 300}

\item{sigma}{Same as in mil_distribution()}

\item{...}{One can pass `kernel_full`, the Gram matrix at instance level for fast computation.}
}
\value{
A list containing the best model and the optimal cost
    parameter.
}
\description{
Function to perform cross validation for the cost used in the
mil_distribution() function. The calculation can be accelerated if
the full gram matrix is provided.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 20,
                            positive_degree = 3,
                            nsample = 10
)
### This takes about 20 minutes.
### foo <- cv_mild(MilData1, n_fold = 2)
###
}
\author{
Yifei Liu
}
