% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{cv_misvm_mip_gurobi}
\alias{cv_misvm_mip_gurobi}
\title{Fit a Cross-Validated MI-SVM model based on full MIP problem}
\usage{
cv_misvm_mip_gurobi(
  y,
  bags,
  X,
  fold_id,
  cost_seq = 2^(-5:15),
  rescale = TRUE,
  verbose = FALSE,
  time_limit = FALSE
)
}
\arguments{
\item{y}{a nx1 numeric vector of bag labels with length equal to the number
of instances (not the number of bags). Must have -1 for negative bags and
+1 for positive bags}

\item{bags}{a nx1 vector specifying which instance belongs to each bag.  Can be
a string, numeric, of factor}

\item{X}{a nxp data.frame of covariates.  Can also supply a matrix.}

\item{fold_id}{a vector indicating which instances belong in each fold for
cross validation.}

\item{cost_seq}{vector of values of c to perform cross vailidation over.  C is
the penalty term on the sum of Xi in the objective function}

\item{rescale}{logical; whether to rescale the input before fitting}

\item{verbose}{whether to message output to the console; default is FALSE.}

\item{time_limit}{FALSE, or a time limit (in seconds) passed to gurobi
parameters. If FALSE, no time limit is given.}
}
\value{
\code{cv_misvm_mip_gurobi()} returns an object of class \code{"misvm"}.
An object of class "misvm" is a list containing at least the following components:
\code{w} a named weight vector matching the number of columns of X
}
\description{
Function to run the gurobi model to train MI-SVM problem.  The hyperparameter
C is chosen through cross-validation.  Cross-validation is done over the
bags and evaluated based on bag AUC.
}
