% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_methods.R
\name{cv_smm_bag}
\alias{cv_smm_bag}
\title{Cross validation for smm_bag method.}
\usage{
cv_smm_bag(
  data,
  n_fold,
  fold_id,
  cost_seq,
  kernel_mild = "rbf",
  class.weights,
  sigma = 0.05
)
}
\arguments{
\item{data}{A MilData object}

\item{n_fold}{The total number of folds}

\item{fold_id}{An optional integer vector representing the fold each bag belongs to}

\item{cost_seq}{A sequence of costs.}

\item{kernel_mild}{Currently can only support 'rbf'}

\item{class.weights}{The class weights, if missing, default to 1:positive/negative bag ratio}

\item{sigma}{The parameter for rbf kernel.}
}
\value{
A list of 4 elements.
}
\description{
Cross validation for smm_bag method.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
## use about 5 minutes
## foo_cv_smm <- cv_smm_bag(data = MilData1, n_fold = 2, cost_seq = 2^(-5:4))
}
\author{
Yifei Liu
}
