% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_samples.R
\name{.compute_cor}
\alias{.compute_cor}
\title{Compute correlations between all features of a data.frame}
\usage{
.compute_cor(data, group_cols)
}
\description{
Compute correlations between all features of a data.frame
}
\author{
Sean Kent
}
\keyword{internal}
