% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.reorder}
\alias{.reorder}
\title{Function to reorder the data by bag label, bag number, and then first data
column}
\usage{
.reorder(y, b, X)
}
\description{
Function to reorder the data by bag label, bag number, and then first data
column
}
\keyword{internal}
