% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_map.R
\name{fit_kernel_feature_map}
\alias{fit_kernel_feature_map}
\title{This flattens the MilData type of data to regular multiple instance data where each instance is a vector}
\usage{
fit_kernel_feature_map(
  data,
  method = c("nystrom", "exact"),
  kernel = c("rbf", "polynomial"),
  output = c("fit", "feature_map", "mean_map"),
  ...
)
}
\arguments{
\item{data}{A MilData object.  Ignored if \code{method} = 'exact' and \code{output} =
"fit".}

\item{...}{}
}
\value{
A data.frame that is ready to be used in \code{MI_SVM()} function.

a function
}
\description{
This flatten the MilData type of data to regular multiple instance data where
each instance is a vector by extracting the features that would be present
in a polynomial kernel with given degree (default = 2).  Each instance_name
has features calculated in the input space based on the underlying kernel
(x'y + c)^d where c = constant and d = degree are input parameters.
TODO: update this to reflect more general methods
}
\details{
TODO: update this
degree degree of the polynomial kernel, default = 2
constant term added to inner product before applying the degree
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
df1 <- build_kernel_features(MilData1, degree = 2)
}
\author{
Sean Kent
}
