% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mild_df.R
\name{generate_mild_df}
\alias{generate_mild_df}
\title{Function to generate mild_df using multivariate t and normal distributions.}
\usage{
generate_mild_df(
  positive_dist = c("mvt", "mvnormal"),
  negative_dist = c("mvnormal", "mvt"),
  remainder_dist = c("mvnormal", "mvt"),
  ncov = 10,
  nimp_pos = 1:5,
  nimp_neg = 1:5,
  nsample = 50,
  ninst = 4,
  nbag = 50,
  positive_mean = rep(0, length(nimp_pos)),
  positive_cov = diag(1, nrow = length(nimp_pos)),
  negative_mean = rep(0, length(nimp_neg)),
  negative_cov = diag(1, nrow = length(nimp_neg)),
  positive_prob = 0.2,
  ...
)
}
\arguments{
\item{positive_dist}{The distribution to be generated for important
covariates of positive instances (default 'mvt').}

\item{negative_dist}{The distribution to be generated for important
covariates of negative instances (default 'mvnormal').}

\item{remainder_dist}{The distribution to be generated for unimportant
covariates (default 'mvnormal').}

\item{ncov}{The number of total covariates (default 10).}

\item{nimp_pos}{An index of important covariates for positve covariates
(default \code{1:5}).}

\item{nimp_neg}{An index of important covariates for negative covariates
(default \code{1:5}).}

\item{nsample}{The number of observations for each instance (default 50).}

\item{ninst}{The number of instances for each bag (default 4).}

\item{nbag}{The number of bags (default 50).}

\item{positive_mean}{The mean vector of important covariates for positive
instances. This should be of same length as \code{nimp_pos}}

\item{positive_cov}{The covariance matrix of important covariates for
positive instances.}

\item{negative_mean}{The mean vector of important covariates for negative
instances, should be of same length as \code{nimp_neg}}

\item{negative_cov}{The covariance matrix of important covariates for
negative instances.}

\item{positive_prob}{A numberic number between 0 and 1 indicating the
probability of an instance being positive.}

\item{...}{Other covariates when using t distribution or using a different
bag labeling scheme, such as:
\itemize{
\item \code{positive_degree} The distribution degree, if \code{positive_dist = 'mvt'}.
\item \code{negative_degree} The distribution degree, if \code{negative_dist = 'mvt'}.
\item \code{remainder_degree} The distribution degree, if \code{remainder_dist = 'mvt'}.
\item \code{positive_bag_prob} The Bernoulli success probability of a bag being a
positive bag. Specify this argument jointly with \code{n_noise_inst} to generate
the label of each bag first.
\item \code{n_noise_inst} The number of negative instances in a positive bag.  This
should be strictly less that \code{ninst}.
}}
}
\value{
A mild_df object.
}
\description{
This function generates multiple instance data (a \code{mild_df} object) where
each row corresponds to a sample from a given instance distribution.
Instances are grouped into bags and the bag labels follow the standard MI
assumption.
}
\details{
To use the function, one needs to provide the total number of covariates
\code{ncov}, to generate, and the index of the important covariates for positive
and negative instances, \code{nimp_pos} and \code{nimp_neg}, respectively. The
distribution of these important covariates will be a multivariate t or
multivariate normal distribution, specified by \code{positive_dist},
\code{negative_dist}. The \code{remainder_dist} specifies the distribution of the
unimportant covariates where their distributions do not differ in positive or
negative instances. \code{nsample} i.i.d draws will be taken from the respective
distributions for each instance. The \code{ninst} parameter specifies the number
of instances in each bag and \code{nbag} gives how many bags in total will be
generated. Hence in total the number of observations is \code{nsample} * \code{nbag} *
\code{ninst}.

The default parameters for the remainder distribution is zero mean and
identity covariance matrix.
}
\examples{
mild_df1 <- generate_mild_df(positive_dist = 'mvt',
                             negative_dist = 'mvnormal',
                             remainder_dist = 'mvnormal',
                             positive_degree = 3)
}
\author{
Yifei Liu
}
