% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{GenerateMilData}
\alias{GenerateMilData}
\title{Function to generate MilData using multivariate t and normal distributions.}
\usage{
GenerateMilData(
  positive_dist = c("mvt", "mvnormal"),
  negative_dist = c("mvnormal", "mvt"),
  remainder_dist = c("mvnormal", "mvt"),
  ncov = 10,
  nimp_pos = 1:5,
  nimp_neg = 1:5,
  nsample = 50,
  ninst = 4,
  nbag = 50,
  positive_mean = rep(0, length(nimp_pos)),
  positive_cov = diag(1, nrow = length(nimp_pos)),
  negative_mean = rep(0, length(nimp_neg)),
  negative_cov = diag(1, nrow = length(nimp_neg)),
  positive_prob = 0.2,
  ...
)
}
\arguments{
\item{positive_dist}{The distribution to be generated for important covariates of positive instances, one of 'mvt' and 'mvnormal'}

\item{negative_dist}{The distribution to be generated for important covariates of negative instances, one of 'mvt' and 'mvnormal'}

\item{remainder_dist}{The distribution to be generated for unimportant covariates, one of 'mvt' and 'mvnormal'.}

\item{ncov}{Number of total covariates.}

\item{nimp_pos}{Index of important covariates for positve covariates.}

\item{nimp_neg}{Index of important covariates for negative covariates.}

\item{nsample}{Number of observations for each instance.}

\item{ninst}{Number of instances for each bag.}

\item{nbag}{Number of bags.}

\item{positive_mean}{The mean vector of important covariates for positive instances, should be of same length as \code{nimp_pos}}

\item{positive_cov}{The covariance matrix of important covariates for positive instances.}

\item{negative_mean}{The mean vector of important covariates for negative instances, should be of same length as \code{nimp_neg}}

\item{negative_cov}{The covariance matrix of important covariates for negative instances.}

\item{positive_prob}{A numberic number between 0 and 1 indicating the probability of an instance being positive.}

\item{...}{Other covariates when using t distribution or using a different bag labeling scheme. If using t distribution, should pass 'positive_degree', 'negative_degree' or 'remainder_degree' in '...' if any of these distributions is specified as 'mvt'. If want to generate the label of each bag first, one needs to specify 'positive_bag_prob' and 'n_noise_inst' in '...'. 'positive_bag_prob' is the Bernoulli success probability of a bag being a positive bag, and 'n_noise_inst' is the number of negative instances in a positive bag, should be strictly less than 'ninst'.}
}
\value{
A MilData object.
}
\description{
This function generates multiple instance data (a MilData object) where each instance corresponds to \code{nsample} i.i.d. observations. To use the function, one needs to provide the total number of covariates \code{ncov}, to generate, and the index of the important covariates for positive and negative instances, \code{nimp_pos} and \code{nimp_neg}, respectively. The distribution of these important covariates will be a multivariate t or multivariate normal distribution, specified by \code{positive_dist}, \code{negative_dist}. The \code{remainder_dist} specifies the distribution of the unimportant covariates where their distributions do not differ in positive or negative instances. The \code{ninst} covariates specifies the number of instances in each bag and \code{nbag} gives how many bags in total will be generated. Hence in total the number of observations is \code{nsample} * \code{nbag} * \code{ninst}.
}
\details{
The default parameters for the remainder distribution is zero mean and identity covariance matrix.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            positive_degree = 3)
}
\author{
Yifei Liu
}
