% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{initialize_instance_selection}
\alias{initialize_instance_selection}
\title{Initialize Instance Selection}
\usage{
initialize_instance_selection(data)
}
\arguments{
\item{data}{A MilData object, potentially generated by MilData().}
}
\value{
list of 3:
'useful_inst_names' includes the names of instances that were selected.
'useful_inst_idx' includes the index of instances that were selected, at the instance level
'yy' includes the bag labels, at the instance level
}
\description{
Use bag_label and instance_name information to initialize the selected
instances. When bag label is 0, select all instances.  When bag label is
1, select the first instance in each bag.
}
\examples{
x = MilData(data.frame('bag_label' = factor(c(1, 1, 0)),
                       'bag_name' = c(rep('bag_1', 2), 'bag_2'),
                       'instance_name' = c('bag_1_inst_1', 'bag_1_inst_2', 'bag_2_inst_1'),
                       'X1' = c(-0.4, 0.5, 2),
                       'instance_label' = c(0, 1, 0)))

initialize_instance_selection(x)

}
\author{
Sean Kent
}
\keyword{internal}
