% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mildsvm.R
\name{kernel_mil}
\alias{kernel_mil}
\title{Function to perform the SMM iteration using full Gram matrix.}
\usage{
kernel_mil(
  kernel_full,
  data_info,
  max.step,
  cost,
  weights,
  sigma,
  yy,
  useful_inst_idx
)
}
\arguments{
\item{kernel_full}{The full Gram matrix, should be of length n_inst * n_inst.}

\item{data_info}{the instance level data information which is a data.frame with 3 columns, 'bag_label', 'bag_name' and 'instance_name'}

\item{max.step}{maximum iteration steps}

\item{cost}{the cost used in SMM}

\item{weights}{Weights of each class}

\item{sigma}{the rbf kernel parameter}

\item{yy}{the response at instance level.}

\item{useful_inst_idx}{a vector specifying which indices are of use.}
}
\value{
A list with several entries.
}
\description{
Internal function to perform SMM iteration using full Gram matrix.
}
\author{
Yifei Liu
}
