% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme.R
\name{kme.default}
\alias{kme.default}
\title{Default method for kme function}
\usage{
\method{kme}{default}(df, df2 = NULL, sigma = 0.05)
}
\arguments{
\item{df}{Data.frame, must have column 'instance_name' which defines the instances.}

\item{df2}{Data.frame}

\item{sigma}{Parameter for 'rbf'}
}
\value{
A matrix.
}
\description{
Default method for kme function.  This function calculates the kernel mean embedding for a given
kernel and a dataset if df2 is NULL, then calculate the kernel mean
embedding matrix of (df, df) otherwise calculate (df, df2)
}
\examples{
x = data.frame('instance_name' = c('inst_1', 'inst_2', 'inst_1'),
               'X1' = c(-0.4, 0.5, 2))
K <- kme(x)
}
\author{
Yifei Liu, Sean Kent
}
