% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme.R
\name{kme.mild_df}
\alias{kme.mild_df}
\title{mild_df method for kme function}
\usage{
\method{kme}{mild_df}(df, df2 = NULL, sigma = 0.05)
}
\arguments{
\item{df}{A mild_df object.}

\item{df2}{Data.frame}

\item{sigma}{Parameter for 'rbf'}
}
\value{
A matrix.
}
\description{
mild_df method for kme function
}
\examples{
mild_df1 <- generate_mild_df(positive_dist = 'mvt',
                             negative_dist = 'mvnormal',
                             remainder_dist = 'mvnormal',
                             nbag = 10,
                             positive_degree = 3)
K <- kme(mild_df1) ## About 10 seconds.
}
\author{
Yifei Liu
}
