% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme.R
\name{kme.MilData}
\alias{kme.MilData}
\title{MilData method for kme function}
\usage{
\method{kme}{MilData}(df, df2 = NULL, sigma = 0.05)
}
\arguments{
\item{df}{A MilData object.}

\item{df2}{Data.frame}

\item{sigma}{Parameter for 'rbf'}
}
\value{
A matrix.
}
\description{
MilData method for kme function
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 10,
                            positive_degree = 3)
K <- kme(MilData1) ## About 10 seconds.
}
\author{
Yifei Liu
}
