% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme.R
\name{kme}
\alias{kme}
\title{Function to calculate the kernel mean embedding matrix for two distributional data with a given kernel}
\usage{
kme(df, df2 = NULL, sigma = 0.05)
}
\arguments{
\item{df}{A data.frame with first column being \code{instance_name} and the rest being features. Or a mild_df object.}

\item{df2}{If \code{df2} is null, the inner embedding of \code{df} and itself will be calculated. Otherwise \code{df2} is a data.frame with first column being \code{instance_name} and the rest being features. Or a mild_df object.}

\item{sigma}{The parameter for rbf kernel.}
}
\value{
A matrix K of number of unique instance labels in df by that in df2.
}
\description{
Function to calculate the kernel mean embedding for to distributional data sets. It uses the empirical approximation for the integral
\deqn{\int_{\mathcal X} \int_{\mathcal Y} K(x, y) d P_X d Q_Y },
for a given kernel \eqn{K(\cdot, \cdot)}. Currently only supports radial basis function kernel for fast computation.
}
\examples{
mild_df1 <- generate_mild_df(positive_dist = 'mvt',
                             negative_dist = 'mvnormal',
                             remainder_dist = 'mvnormal',
                             nbag = 10,
                             positive_degree = 3)
K <- kme(mild_df1)
}
\author{
Yifei Liu, Sean Kent
}
