% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme.R
\name{linear_kernel_matrix}
\alias{linear_kernel_matrix}
\title{Compute the Linear Kernel from two matrices}
\usage{
linear_kernel_matrix(x, y)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix, must have same number of columns as \code{x}}
}
\description{
Compute the Linear Kernel from two matrices.  Follows
very closesly to kernlab::kernelMatrix.vanilla but omits unnecessary
checks that slow down the computation.
}
\author{
Sean Kent
}
