% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{MI_SVM}
\alias{MI_SVM}
\title{MI-SVM algorithm implementation in R}
\usage{
MI_SVM(
  data,
  cost,
  kernel = "radial",
  max.step = 500,
  type = "C-classification"
)
}
\arguments{
\item{data}{A data.frame whose first three columns are \code{bag_label}, \code{bag_name} and \code{instance_name}.}

\item{cost}{The cost parameter to be fed to \code{e1071::svm}.}

\item{kernel}{The kernel function to be used for \code{e1071::svm}.}

\item{max.step}{Maximum steps of iteration for the iterative SVM methods.}

\item{type}{type that to be used for \code{e1071::svm}.}
}
\value{
An object of class 'MI_SVM'
}
\description{
This function implements the MI-SVM algorithm proposed by Andrews et al (2003)
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
df1 <- build_instance_feature(MilData1, seq(0.05, 0.95, length.out = 10))
mdl <- MI_SVM(data = df1, cost = 1, kernel = 'radial')
}
\author{
Yifei Liu
}
