% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mil_distribution.R
\name{mil_distribution}
\alias{mil_distribution}
\title{Function to implement the iterative multiple instance learning with distributional data algorithm.}
\usage{
mil_distribution(data, cost, max.step = 500, sigma = 0.05, ...)
}
\arguments{
\item{data}{A MilData object, potentially generated by MilData().}

\item{cost}{The cost for smm.}

\item{max.step}{The total number of iterations.}

\item{sigma}{The parameter for the rbf kernel.}

\item{...}{One can pass `kernel_full`, the Gram matrix at instance level for fast computation.}
}
\value{
A mild object which contains the results.
}
\description{
Workhorse for the package. This implements the algorithm that
iteratively find the representative positive instances and use smm
to get the model.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 10,
                            positive_degree = 3
                           )
foo <- mil_distribution(data = MilData1, cost = 1) ## uses about 10 seconds.
}
\author{
Yifei Liu
}
