% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_methods.R
\name{mil_with_feature}
\alias{mil_with_feature}
\title{Use MI_SVM on flattened dataset}
\usage{
mil_with_feature(
  data,
  kernel = "radial",
  cost = 1,
  class.weights = NULL,
  sigma = 0.05,
  type = "C-classification",
  qtls = seq(0.05, 0.95, length.out = 10),
  max.step = 500
)
}
\arguments{
\item{data}{A MilData object}

\item{kernel}{kernel to be used by MI_SVM}

\item{cost}{Cost to be used by MI_SVM}

\item{class.weights}{class weights to be used by MI_SVM}

\item{sigma}{sigma to be used by MI_SVM}

\item{type}{type to be used by MI_SVM}

\item{qtls}{quantiles to be used by build_instance_feature}

\item{max.step}{maximum steps to be used by MI_SVM}
}
\value{
The result returned by MI_SVM
}
\description{
This function just combines build_instance_feature() and MI_SVM()
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
foo <- mil_with_feature(data = MilData1)
}
\author{
Yifei Liu
}
