% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mild_df.R
\name{mild_df}
\alias{mild_df}
\title{Build a MILD data frame}
\usage{
mild_df(
  bag_label = character(),
  bag_name = character(),
  instance_name = character(),
  instance_label = NULL,
  ...
)
}
\arguments{
\item{bag_label}{A character, factor, or numeric vector}

\item{bag_name}{A character or factor vector}

\item{instance_name}{A character or factor vector}

\item{instance_label}{A character, factor, or numeric vector, or NULL}
}
\value{
A 'mild_df' object.
}
\description{
\code{mild_df()} constructs a data frame that corresponds to Multiple Instance
Learning with Distributional Instances (MILD) data.  A 'mild_df' object must
have three special columns:
\itemize{
\item 'bag_label', determines the label of each bag, typically from c(0, 1)
\item 'bag_name', character or factor that specifies the bag that each sample
belongs to.
\item 'instance_name', character or factor that specifies the instance that
each sample belongs to.
}
}
\details{
We refer to the rows of a 'mild_df' as \emph{samples}, since they are
thought of as draws from the distribution that determines each instance.
Each instance is contained in a bag, with a corresponding label.  Instance
labels can be provided, but they will be pulled in as an attribute.
}
\examples{
mild_df('bag_label' = factor(c(1, 1, 0)),
        'bag_name' = c(rep('bag_1', 2), 'bag_2'),
        'instance_name' = c('bag_1_inst_1', 'bag_1_inst_2', 'bag_2_inst_1'),
        'X1' = c(-0.4, 0.5, 2),
        'instance_label' = c(0, 1, 0))

}
\author{
Yifei Liu, Sean Kent
}
