% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{MilData}
\alias{MilData}
\title{Helper function to generate MilData object}
\usage{
MilData(x = data.frame())
}
\arguments{
\item{x}{A data.frame that to be used to create MilData object.}
}
\value{
A MilData object.
}
\description{
Helper function to generate MilData object. Changes the first three columns to the required column names ('bag_label', 'bag_name', 'instance_name'). If the original `bag_label` is a factor, will use the first level as negative label and code it as numeric '0' and use the second level as positive label and code it as numeric '1'. Levels >= 3 not supported and will raise an error.
}
\details{
The column name for instance label should always be `instance_label` otherwise will be taken as features of x.
}
\examples{
x = data.frame('bag_LABEL' = factor(c(1, 1, 0)),
              'bag_name' = c(rep('bag_1', 2), 'bag_2'),
              'instance_name' = c('bag_1_inst_1', 'bag_1_inst_2', 'bag_2_inst_1'),
              'X1' = c(-0.4, 0.5, 2),
              'instance_label' = c(0, 1, 0))
MilData(x)
}
\author{
Yifei Liu
}
