% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{misvm_mip_gurobi}
\alias{misvm_mip_gurobi}
\title{Fit MI-SVM model based on full MIP problem}
\usage{
misvm_mip_gurobi(
  y,
  bags,
  X,
  c,
  rescale = TRUE,
  weights = NULL,
  verbose = FALSE,
  time_limit = FALSE
)
}
\arguments{
\item{y}{a nx1 numeric vector of bag labels with length equal to the number
of instances (not the number of bags). Must have -1 for negative bags and
+1 for positive bags}

\item{bags}{a nx1 vector specifying which instance belongs to each bag.  Can be
a string, numeric, of factor}

\item{X}{a nxp data.frame of covariates.  Can also supply a matrix.}

\item{c}{scalar indicating the penalty term on the sum of $\xi$ in the
objective function}

\item{rescale}{logical; whether to rescale the input before fitting}

\item{weights}{named vector to control the weight of the cost parameter
for each possible y value.  Weights multiply against the cost vector.}

\item{verbose}{whether to message output to the console; default is FALSE.}

\item{time_limit}{the time limit for each iteration of cross validation.}
}
\value{
\code{misvm_mip_gurobi()} returns an object of class \code{"misvm"}.
An object of class "misvm" is a list containing at least the following
components:
\itemize{
\item \code{w} a named weight vector matching the number of columns of X
TODO: fill out the rest of this...
}
}
\description{
Function to run the Gurobi model to train MI-SVM problem based on the full
specification of the Mixed Integer Programming (MIP) problem.
}
