% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{misvm_mip_model}
\alias{misvm_mip_model}
\title{INTERNAL Create optimization model for MI-SVM problem}
\usage{
misvm_mip_model(y, bags, X, c, weights = NULL)
}
\arguments{
\item{y}{a nx1 numeric vector of bag labels with length equal to the number
of instances (not the number of bags). Must have -1 for negative bags and
+1 for positive bags}

\item{bags}{a nx1 vector specifying which instance belongs to each bag.  Can be
a string, numeric, of factor}

\item{X}{an nxp data.frame of covariates.  Can also supply a matrix.}

\item{c}{scalar indicating the penalty term on the sum of Xi in the
objective function}

\item{weights}{named vector to control the weight of the cost parameter
for each possible y value.  Weights multiply against the cost vector.}
}
\value{
a model that can be passed to \code{gurobi::gurobi} that contains the MIQP
problem defined by MI-SVM in Andrews et al. (2003)
}
\description{
Internal function to build an optimization model (that can be passed to
\code{gurobi::gurobi}) based on the MI-SVM problem.
}
\author{
Sean Kent
}
