% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{misvm_qpheuristic_fit}
\alias{misvm_qpheuristic_fit}
\title{INTERNAL Fit MI-SVM model based on heuristic method and QP optimization}
\usage{
misvm_qpheuristic_fit(
  y,
  bags,
  X,
  c,
  rescale = TRUE,
  weights = NULL,
  verbose = FALSE,
  time_limit = FALSE,
  max_step = 500
)
}
\arguments{
\item{y}{a nx1 numeric vector of bag labels with length equal to the number
of instances (not the number of bags). Must have -1 for negative bags and
+1 for positive bags}

\item{bags}{a nx1 vector specifying which instance belongs to each bag.  Can be
a string, numeric, of factor}

\item{X}{an nxp data.frame of covariates.  Can also supply a matrix.}

\item{c}{scalar indicating the penalty term on the sum of Xi in the
objective function}

\item{rescale}{logical; whether to rescale the input before fitting}

\item{weights}{named vector to control the weight of the cost parameter
for each possible y value.  Weights multiply against the cost vector.}

\item{verbose}{whether to message output to the console; default is FALSE.}

\item{time_limit}{FALSE, or a time limit (in seconds) passed to gurobi
parameters. If FALSE, no time limit is given.}
}
\value{
\code{misvm_qpheuristic_fit()} returns an object of class \code{"misvm"}.
An object of class "misvm" is a list containing at least the following
components:
\itemize{
\item \code{model}: a list with components:
\itemize{
\item \code{w}: weights to apply to the predictors to make the classifier
\item \code{b}: intercept term used to make the classifier
\item \code{xi}: slack variables returned from the optimization
\item \code{status}: solution status from \code{gurobi::gurobi}
\item \code{itercount}: itercount from \code{gurobi::gurobi}
\item \code{baritercount}: baritercount from \code{gurobi::gurobi}
\item \code{objval}: value of the objective at the solution
\item \code{c}: value of the cost parameter used in solving the optimization
\item \code{n_selections}: the number of selections used in fitting
}
\item \code{representative_inst}: NULL, TODO: mesh with other misvm method
\item \code{traindata}: NULL, TODO: mesh with other misvm method
\item \code{useful_inst_idx}: NULL, TODO: mesh with other misvm method
}
}
\description{
Function to train an MI-SVM classifier based on the heuristic method proposed
by Andrews et al. (2003) where each step is solved with the Quadratic
Programming (QP) problem with correct constraints.  This can only be solved
with a linear space (no kernel methods are available).  The optimization
problem is solved using the \code{gurobi} R package through the Gurobi backend.
}
\author{
Sean Kent
}
\keyword{internal}
