% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mil_distribution.R
\name{new_mild}
\alias{new_mild}
\title{Function to construct mild object}
\usage{
new_mild(x = list())
}
\arguments{
\item{x}{A list}
}
\value{
A mild object.
}
\description{
Function to construct mild object. `x` should be a list with components `model` (an smm object), `total_step` (a integer), `representative_inst` (a matrix of two columns), `traindata` (a data.frame with the first column being 'instance_name')
}
\examples{
new_mild(list())
}
\author{
Yifei Liu
}
