% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mild_data.R
\name{new_MilData}
\alias{new_MilData}
\title{Constructor function for MilData object}
\usage{
new_MilData(x = data.frame())
}
\arguments{
\item{x}{A data.frame which contains 'bag_label', 'bag_name', 'instance_name' as the first three columns and the rest as features.}
}
\value{
A MilData object (S3) which is built on data.frame
}
\description{
Constructor function for MilData object which gives a more convenient way of handling multiple instance data. If x contains a column for 'instance_label', then this function will add \code{instance_label} in the object attributes but will remove it from the data.frame.
}
\examples{
new_MilData(x = data.frame('bag_label' = c(1, 1),
                           'bag_name' = rep('bag_1', 2),
                           'instance_name' = c('bag_1_inst_1', 'bag_1_inst_2'),
                           'X1' = c(-0.4, 0.5))
                           )
new_MilData(x = data.frame('bag_label' = c(1, 1),
                           'bag_name' = rep('bag_1', 2),
                           'instance_name' = c('bag_1_inst_1', 'bag_1_inst_2'),
                           'X1' = c(-0.4, 0.5),
                           'instance_label' = c(0, 1))
                           )
}
\author{
Yifei Liu
}
