% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{new_smm}
\alias{new_smm}
\title{Create a new smm object}
\usage{
new_smm(x = list())
}
\arguments{
\item{x}{A list}
}
\value{
An smm object
}
\description{
Create a new smm object
}
\examples{

new_smm(list())
}
\author{
Yifei Liu
}
