% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ordmvnorm}
\alias{ordmvnorm}
\title{Sample ordinal MIL data using mvnorm}
\format{
A data frame with 1000 rows and 8 variables:
\describe{
\item{bag_label}{outcome label at the bag level. This is the maximum of the \code{inst_label} for each bag}
\item{bag_name}{indicator of each bag}
\item{inst_label}{outcome label at the instance level}
\item{V1}{Variable with mean equal to \code{2 * inst_label}}
\item{V2}{Variable with mean equal to \code{-1 * inst_label}}
\item{V3}{Variable with mean equal to \code{1 * inst_label}}
\item{V4}{Variable with mean 0, essentially noise}
\item{V5}{Variable with mean 0, essentially noise}
}
}
\usage{
ordmvnorm
}
\description{
A data set that demonstrates the ordinal multiple-instance learning
structure with feature columns randomly sampled from a multivariate normal
distribution.
}
\keyword{datasets}
