% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mil_distribution.R
\name{predict.mild}
\alias{predict.mild}
\title{Prediction function for mild objects}
\usage{
predict.mild(object, ...)
}
\arguments{
\item{object}{A mild object}

\item{...}{should pass 'newdata', the new dataset to be predict on, optionally 'GramMatrix' to accelarate the speed.}
}
\value{
A data.frame with the same number of rows as in newdata and with four columns, 'bag_name', 'instance_name', 'bag_label_pred', 'instance_score'
}
\description{
Prediction function for mild objects.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 10,
                            positive_degree = 3,
                            nsample = 20
)
foo <- mil_distribution(data = MilData1, cost = 1) ## uses about 10 seconds.
predictions_mild <- predict(object = foo, newdata = MilData1)
}
\author{
Yifei Liu
}
