% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{predict.misvm}
\alias{predict.misvm}
\title{Predict method for 'misvm' object}
\usage{
\method{predict}{misvm}(object, newX, type = c("prediction", "score"))
}
\arguments{
\item{object}{an object of class misvm}

\item{newX}{matrix to predict from.  Needs to have the same number of
columns as the X that trained the misvm object}

\item{type}{if 'prediction', return predicted values with threshold of 0 as
-1 or +1.  If 'score', return the raw predicted scores.}
}
\value{
vector of length \code{nrow(newX)}.  If 'prediction', return predicted
values with threshold of 0 as -1 or +1.  If 'scores', return the raw
predicted scores.
}
\description{
Predict method for 'misvm' object
}
\author{
Sean Kent
}
