% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMM.R
\name{predict.smm}
\alias{predict.smm}
\title{Prediction function for smm objects (the return from SMM())}
\usage{
\method{predict}{smm}(object, ...)
}
\arguments{
\item{object}{An smm object.}

\item{...}{`newdata`, new data whose first column should be `instance_name`. `traindata`, training data used to train res. The first column should be `instance_name`. `kernel_mild`, currently only support 'rbf'}
}
\value{
A vector of length of the instances in newdata.
}
\description{
This function gives the prediction of an smm object in the scale of score. A higher score indicates the observation is more likely to be of class `1`.
}
\examples{
x = data.frame('bag_LABEL' = factor(c(1, 1, 0)),
               'bag_name' = c(rep('bag_1', 2), 'bag_2'),
               'instance_name' = c('bag_1_inst_1', 'bag_1_inst_2', 'bag_2_inst_1'),
               'X1' = c(-0.4, 0.5, 2),
               'instance_label' = c(0, 1, 0)
)
mdl <- SMM(df = MilData(x), cost = 10, sigma = 0.05)
x_inst <- data.frame('instance_name' = c('bag_1_inst_1', 'bag_1_inst_2', 'bag_2_inst_1'),
               'X1' = c(-0.4, 0.5, 2)
               )
predictions <- predict(object = mdl, kernel_mild = 'rbf', newdata = x_inst, traindata = x_inst)
}
\author{
Yifei Liu
}
