% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_methods.R
\name{predict.smmBag}
\alias{predict.smmBag}
\title{Predict method for smmBag objects}
\usage{
predict.smmBag(object, ...)
}
\arguments{
\item{object}{An smmBag object}

\item{...}{newdata, GramMatrix for faster calculation.}
}
\value{
An list
}
\description{
Predict method for smmBag objects
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 20,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
foo_smm <- smm_bag(data = MilData1)
foo_predict <- predict(foo_smm, newdata = MilData1)
}
\author{
Yifei Liu
}
