% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfm_exact.R
\name{predict_exact}
\alias{predict_exact}
\alias{predict_exact.default}
\alias{predict_exact.MilData}
\title{Predict feature map using exact features}
\usage{
predict_exact(object, newx)

\method{predict_exact}{default}(object, newx)

\method{predict_exact}{MilData}(object, newx)
}
\arguments{
\item{object}{a object from a call to \code{fit_kernel_feature_map()} with
\code{method} = 'exact'.}

\item{newx}{An object containing features for prediction. Usually a data.frame
or matrix.}
}
\value{
a matrix of features that use the exact kernel feature map for each
observation in \code{newx}.
}
\description{
TODO: add description
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: For use on objects of class \code{data.frame} or \code{matrix}.

\item \code{MilData}: Ignore the information columns with 'bag_label',
'bag_name', and 'instance_name' when calculating kernel approximation.
These columns are re-appended upon prediction.
}}

\examples{
df <- data.frame(
  X1 = c(2,   3,   4,   5,   6, 7, 8),
  X2 = c(1, 1.2, 1.3, 1.4, 1.1, 7, 1),
  X3 = rnorm(7)
)

fit <- fit_kernel_feature_map(df, method = "exact", kernel = "polynomial", degree = 2, const = 1)
fm <- predict_exact(fit, df)

}
\author{
Sean Kent
}
