% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_map.R
\name{predict_nystrom}
\alias{predict_nystrom}
\title{Predict feature map based on Nystrom approximation}
\usage{
predict_nystrom(object, newx)
}
\arguments{
\item{object}{a object from a call to \code{fit_nystrom()}}

\item{newx}{An object containing features for prediction. Usually a data.frame
or matrix. Must contain the same columns as \code{df} from the fitting.}
}
\value{
a matrix of features that approximate the fitted kernel for each
observation in \code{newx}.
}
\description{
TODO: add description
}
\examples{
df <- data.frame(
  X1 = c(2,   3,   4,   5,   6, 7, 8),
  X2 = c(1, 1.2, 1.3, 1.4, 1.1, 7, 1),
  X3 = rnorm(7)
)

fit <- fit_nystrom(df, m = 7, r = 6, kernel = "rbf", sigma = 0.05)
fm <- predict_nystrom(fit, df)

}
\author{
Sean Kent
}
