% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme.R
\name{rbf_kernel_matrix}
\alias{rbf_kernel_matrix}
\title{Compute the Radial Basis Kernel from two matrices}
\usage{
rbf_kernel_matrix(sigma, x, y)
}
\arguments{
\item{sigma}{Parameter for the radial basis function that controls the width}

\item{x}{matrix}

\item{y}{matrix, must have same number of columns as `x`}
}
\description{
Compute the Radial Basis Kernel from two matrices.  Follows
very closesly to kernlab::kernelMatrix.rbfkernel but omits unnecessary
checks that slow down the computation.
}
\author{
Sean Kent
}
