% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{select_cv_folds}
\alias{select_cv_folds}
\title{Select Cross Validation folds from MilData}
\usage{
select_cv_folds(data, n_fold, fold_id)
}
\arguments{
\item{data}{A MilData object, potentially generated by MilData().}

\item{n_fold}{the number of folds for cross validation}

\item{fold_id}{A user specified fold_id. Suggested not to specify
this.}
}
\value{
a list with the following components:
\itemize{
\item \code{n_fold} the number of folds for cross validatoin
\item \code{fold_id} if fold_id is missing in input, returns the folds calculated by
splitting the positive and negative bags separately.
}
}
\description{
Uses the bag information from a \code{data} object to generate folds for use in
cross validation.
}
\author{
Yifei Liu, Sean Kent
}
\keyword{internal}
