% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{select_cv_folds2}
\alias{select_cv_folds2}
\title{Function that will eventually supercede select_cv_folds because it supports
the same variables at the new function interface.}
\usage{
select_cv_folds2(y, bags, n_fold, fold_id)
}
\description{
Function that will eventually supercede select_cv_folds because it supports
the same variables at the new function interface.
}
\keyword{internal}
