% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMM.R
\name{SMM.MilData}
\alias{SMM.MilData}
\title{MilData method for SMM function}
\usage{
\method{SMM}{MilData}(
  df,
  kernel_mild = "rbf",
  cost = 1,
  class.weights = NULL,
  sigma = 0.05,
  y = NULL
)
}
\arguments{
\item{df}{A data.frame where the first two columns being `instance_label` and `instance_name`, or a MilData object. `df` can be NULL if kernel_mild is a matrix and y is not NULL.}

\item{kernel_mild}{The kernel to be used, current only supports 'rbf'; or a kernal matrix.}

\item{cost}{The cost for SMM, larger cost gives larger tolerance to errors.}

\item{class.weights}{A named vector where the names should be the same as the unique values of the label. For the use of imbalance data. Will be passed to the `ksvm` function in `kernlab` package. Notice that a class weight 2:1 will gives different results from 1:0.5 because they will be directly multiplied to the cost.}

\item{sigma}{The parameter for the rbf kernel.}

\item{y}{The instance label. If omitted, will use that from df.}
}
\value{
An `smm` object.
}
\description{
MilData method for SMM function
}
\examples{

x = data.frame('bag_LABEL' = factor(c(1, 1, 0)),
               'bag_name' = c(rep('bag_1', 2), 'bag_2'),
               'instance_name' = c('bag_1_inst_1', 'bag_1_inst_2', 'bag_2_inst_1'),
               'X1' = c(-0.4, 0.5, 2),
               'instance_label' = c(0, 1, 0)
)
mdl <- SMM(df = MilData(x), cost = 10)
}
\author{
Yifei Liu
}
