% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_methods.R
\name{smm_bag}
\alias{smm_bag}
\title{Function to perform Support Measure Machines for MilData objects.}
\usage{
smm_bag(data, kernel_mild = "rbf", cost = 1, class.weights, sigma = 0.05, ...)
}
\arguments{
\item{data}{A MilData object.}

\item{kernel_mild}{Same as \code{SMM()}}

\item{cost}{Same as \code{SMM()}}

\item{class.weights}{Same as \code{SMM()}}

\item{sigma}{Same as \code{SMM()}}

\item{...}{If kernel_mild is a matrix, need to pass instance level label \code{y}.}
}
\value{
A list
}
\description{
This function solves the mil with distributional data
classification problem using the naive SMM method by regarding bag
lable as instance label.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
foo_smm <- smm_bag(data = MilData1)
}
\author{
Yifei Liu
}
