% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_samples.R
\name{summarize_samples}
\alias{summarize_samples}
\alias{summarize_samples.default}
\alias{summarize_samples.MilData}
\title{Summarize data across functions}
\usage{
\method{summarize_samples}{default}(data, group_cols, .fns = list(mean = mean), cor = FALSE, ...)

\method{summarize_samples}{MilData}(data, ...)
}
\arguments{
\item{data}{a data.frame, 'MilData' object, or similar of data to summarize}

\item{group_cols}{character vector of column(s) that describe groups to
summarize across}

\item{.fns}{list of functions (default list(mean = mean)).}

\item{cor}{logical (default FALSE), whether to include correlations between
all features in the summarization}
}
\value{
a 'tibble' with summarized data.  There will be one row for each set
of distinct groups specified by \code{group_cols}. There will be one column for
each of the \code{group_cols} plus \code{length(.fns)} columns for each of the
features in \code{data} plus correlation columns if specified.
}
\description{
Summarize a numeric data frame based on specified grouping columns and a list
of functions.  This is useful in summarizing a 'MilData' object from the
sample level to the instance level.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Method for data.frame-like objects.

\item \code{MilData}: Method for 'MilData' objects.
}}

\examples{
fns <- list(mean = mean, sd = sd)
summarize_samples(mtcars, group_cols = c("cyl", "gear"), .fns = fns)
summarize_samples(mtcars, group_cols = c("cyl", "gear"), .fns = fns, cor = TRUE)

}
\author{
Sean Kent
}
