% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{validate_MilData}
\alias{validate_MilData}
\title{Validation function for MilData object}
\usage{
validate_MilData(x)
}
\arguments{
\item{x}{A data.frame. x should have the first three column names representing 'bag_label', 'bag_name' and 'instance_name'. The rest should all be features. Therefore, the bag_label of the same bag_name should be the same.}
}
\value{
A MilData object if it checks out, or an error message
}
\description{
x should have the first three column names representing 'bag_label', 'bag_name' and 'instance_name'. The rest should all be features. Therefore, the bag_label of the same bag_name should be the same.
}
\examples{
x = data.frame('bag_label' = c(1, 1, 0),
              'bag_name' = c(rep('bag_1', 2), 'bag_2'),
              'instance_name' = c('bag_1_inst_1', 'bag_1_inst_2', 'bag_2_inst_1'),
              'X1' = c(-0.4, 0.5, 2),
              'instance_label' = c(0, 1, 0))
validate_MilData(x)
}
\author{
Yifei Liu
}
