% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{classify_bags}
\alias{classify_bags}
\title{Classify y from bags}
\usage{
classify_bags(y, bags, condense = TRUE)
}
\arguments{
\item{y}{a numeric vector of bag labels with length equal to the number of
instances (not the number of bags). Must have -1 for negative bags and
+1 for positive bags}

\item{bags}{a vector specifying which instance belongs to each bag.  Can be
a string, numeric, of factor}

\item{condense}{whether to return classification at the level of unique bags or not}
}
\value{
a named vector of length \code{length(unique(b))} which gives the
classification for each bag.  Names come from \code{bags}.
}
\description{
Classify y from bags
}
\author{
Sean Kent
}
