% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{cv_MI_SVM}
\alias{cv_MI_SVM}
\title{Cross-validation function for MI_SVM}
\usage{
cv_MI_SVM(
  data,
  n_fold,
  fold_id,
  cost_seq,
  kernel = "radial",
  max.step = 500,
  type = "C-classification"
)
}
\arguments{
\item{data}{A data.frame whose first three columns are \code{bag_label}, \code{bag_name} and \code{instance_name}.}

\item{n_fold}{The number of folds.}

\item{fold_id}{A user specified fold_id. Suggested not to specify
this.}

\item{cost_seq}{The cost sequence, default to be 2^(-2:2).}

\item{kernel}{The kernel function to be used for \code{e1071::svm}.}

\item{max.step}{Maximum steps of iteration for the iterative SVM methods.}

\item{type}{type that to be used for \code{e1071::svm}.}
}
\value{
A list that contains best model, optimal cost value, the AUCs, and the cost sequence.
}
\description{
Cross-validation function for MI_SVM.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
df1 <- build_instance_feature(MilData1, seq(0.05, 0.95, length.out = 10))
foo_cv <- cv_MI_SVM(data = df1, n_fold = 3, cost_seq = 2^(-2:2))
}
\author{
Yifei Liu
}
