% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mil_distribution.R
\name{initialize_instance_selection}
\alias{initialize_instance_selection}
\title{Initialize Instance Selection}
\usage{
initialize_instance_selection(data)
}
\arguments{
\item{data}{A MilData object, potentially generated by MilData().}
}
\value{
list of 3:
'useful_inst_names' includes the names of instances that were selected.
'useful_inst_idx' includes the index of instances that were selected, at the instance level
'yy' includes the bag labels, at the instance level
}
\description{
Use bag_label and instance_name information to initialize the selected
instances. When bag label is 0, select all instances.  When bag label is
1, select the first instance in each bag.
}
\author{
Sean Kent
}
