% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfm_exact.R
\name{kfm_exact}
\alias{kfm_exact}
\title{Create an exact kernel feature map}
\usage{
kfm_exact(kernel = "polynomial", degree = 2, const = 1)
}
\arguments{
\item{kernel}{a character determining the kernel to use.  Currently, only
'radial' is implemented.}

\item{degree}{numeric (default 2), provides the degree for \code{kernel} =
'polynomial'}

\item{const}{numeric (default 1), constant term for \code{kernel} = 'polynomial'}
}
\value{
an object of class 'kfm_exact' with the following components,
returned from the inputs:
\itemize{
\item \code{kernel}
\item \code{degree}
\item \code{const}
}
}
\description{
For some kernels, it is possible to create the exact features from given
data. This function stores the information needed to build those exact
features.
}
\details{
Currently, the following kernels are supported:
\itemize{
\item 'polynomial', with \code{degree} = d and \code{const} = c
}
}
\examples{
df <- data.frame(
  X1 = c(2,   3,   4,   5,   6, 7, 8),
  X2 = c(1, 1.2, 1.3, 1.4, 1.1, 7, 1),
  X3 = rnorm(7)
)

fit <- kfm_exact(kernel = "polynomial", degree = 2, const = 1)
fm <- build_fm(fit, df)

}
\author{
Sean Kent
}
