% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkfm_functions.R
\name{kfm_nystrom}
\alias{kfm_nystrom}
\alias{kfm_nystrom.default}
\alias{kfm_nystrom.MilData}
\alias{predict_kfm_nystrom.default}
\alias{predict_kfm_nystrom.MilData}
\title{Fit a Nystrom method kernel approximation}
\usage{
kfm_nystrom(df, m, r, kernel, ...)

\method{kfm_nystrom}{default}(df, m = nrow(df), r = m, kernel = "rbf", sampling = "random", ...)

\method{kfm_nystrom}{MilData}(df, m = nrow(df), r = m, kernel = "rbf", sampling = "random", ...)

\method{predict_kfm_nystrom}{default}(object, newx)

\method{predict_kfm_nystrom}{MilData}(object, newx)
}
\arguments{
\item{df}{An object containing features for training.  Usually a data.frame
or matrix.}

\item{m}{the number of examples from \code{df} to sample in fitting.}

\item{r}{the rank of matrix approximation to use. Must be less than or equal
to \code{m}, the default.}

\item{kernel}{a character determining the kernel to use.  Currently, only
'rbf' is implemented.}

\item{...}{additional parameters needed for the kernels.  See details.}

\item{sampling}{determines how to sample instances.  Default it 'random'.
For kfm_nystrom.MilData, can specify \code{sampling} = 'stratified' to ensure
that samples are chosen evenly from bags and instances.  \code{sampling} can
also be a numeric vector of length \code{m} of pre-determined samples.}
}
\value{
an S3 object with the following components
\itemize{
\item \code{df_sub} the sub-sampled version of \code{df}
\item \code{dv} pre-multiplication matrix which contains information on the
eigenvalues and eigenvectors of \code{df_sub}
\item \code{method} 'nystrom'
\item \code{kernel} the input parameter \code{kernel}
\item \code{kernel_params} parameters passed to \code{...}
}
}
\description{
TODO: add description
}
\details{
For the \code{...} argument, the additional parameters depend on which
kernel is used:
\itemize{
\item For \code{kernel} = 'rbf', specify \code{sigma} to define kernel bandwidth.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: For use on objects of class \code{data.frame} or \code{matrix}.

\item \code{MilData}: Ignore the information columns with 'bag_label',
'bag_name', and 'instance_name' when calculating kernel approximation.
These columns are re-appended upon prediction.

\item \code{default}: For use on objects of class \code{data.frame} or \code{matrix}.

\item \code{MilData}: Ignore the information columns with 'bag_label',
'bag_name', and 'instance_name' when calculating kernel approximation.
These columns are re-appended upon prediction.
}}

\examples{
df <- data.frame(
  X1 = c(2,   3,   4,   5,   6, 7, 8),
  X2 = c(1, 1.2, 1.3, 1.4, 1.1, 7, 1),
  X3 = rnorm(7)
)

fit <- kfm_nystrom(df, m = 7, r = 6, kernel = "rbf", sigma = 0.05)
fm <- predict_kfm_nystrom(fit, df)

}
\author{
Sean Kent
}
