% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mild.R
\name{mild}
\alias{mild}
\title{Create a mild object}
\usage{
mild(bag_label, bag_name, instance_name)
}
\arguments{
\item{bag_label}{the bag label or response, recorded as 0 = negative, 1 =
positive}

\item{bag_name}{a unique bag identifier for each instance.}

\item{instance_name}{a unique instance identifier for each sample.}
}
\value{
An object of class \code{mild}.  Currently, no methods are implemented for
this
}
\description{
Create a \code{mild} object, usually used as a response variable in a model
formula.
}
\examples{

mild_data <- generate_mild_df(
  positive_dist = 'mvt',
  negative_dist = 'mvnormal',
  remainder_dist = 'mvnormal',
  nbag = 50,
  nsample = 20,
  positive_degree = 3,
  positive_prob = 0.15,
  positive_mean = rep(0, 5)
)
with(mild_data, mild(bag_label, bag_name, instance_name))
df <- get_all_vars(mild(bag_label, bag_name) ~ X1 + X2, data = mild_data)
head(df)

}
\author{
Sean Kent
}
