% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mildsvm.R
\name{mildsvm}
\alias{mildsvm}
\title{Function to perform cross validation for the cost used in mil_distribution() function.}
\usage{
mildsvm(
  data,
  cost = 1,
  method = "mip",
  weights = TRUE,
  time_limit = 150,
  kfm_fun = "kfm_nystrom",
  ...
)
}
\arguments{
\item{data}{The dataset as a MilData object.}

\item{cost}{the cost parameter in the model; default = 1. TODO: make this better}

\item{method}{the method to use in fitting. If "mip", uses the full mildsvm
mixed integer program.}

\item{weights}{if TRUE, will adjust the SVM cost to account for class imbalance
if FALSE or NULL, no adjustment is made. Optionally, one can pass pre-specified
weights like \code{c("0" = 2, "1" = 1)} that get used.  Default is \code{weights = TRUE}.}

\item{time_limit}{the time limit for each iteration of cross validation.}

\item{kfm_fun}{a kernel feature map fitting function required when method = "mip"}

\item{...}{additional arguments passed to the \code{kfm_fun}}
}
\value{
An object of class \code{mildsvm} fit to the data.
}
\description{
Function to perform cross validation for the cost used in the
mil_distribution() function. The calculation can be accelerated if
the full gram matrix is provided.
TODO: update parameters, return, etc.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 15,
                            positive_degree = 3,
                            nsample = 20
)
foo <- mildsvm(MilData1, method = "mip", m = 10)

}
\author{
Sean Kent
}
