% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{misvm}
\alias{misvm}
\alias{misvm.formula}
\alias{misvm.default}
\title{fit MI-SVM model to the data}
\usage{
\method{misvm}{formula}(
  formula,
  data,
  cost,
  kernel,
  method = c("heuristic", "mip"),
  ...,
  weights = TRUE
)

\method{misvm}{default}(
  x,
  y,
  bags,
  cost,
  kernel,
  method = c("heuristic", "mip"),
  ...,
  weights = TRUE
)
}
\description{
Description
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: Method for passing formula

\item \code{default}: Method for data.frame-like objects
}}

