% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm.R
\name{predict.MI_SVM}
\alias{predict.MI_SVM}
\title{Prediction function for MI_SVM object}
\usage{
\method{predict}{MI_SVM}(object, ...)
}
\arguments{
\item{object}{The return from MI_SVM()}

\item{...}{Should include \code{newdata} to represent the newdata to be predicting on, and optionally \code{true_bag_info} data.frame that contains 'bag_label' and 'bag_name' for AUC and ROC calculation.}
}
\value{
A list which contains a bag level prediction \code{bag_level_prediction} and an instance level prediction \code{instance_level_prediction}. If \code{true_bag_label} is fed to arguments, will also return \code{ROC} and \code{AUC} as list elements.
}
\description{
Predictionn function for MI_SVM object.
}
\examples{
MilData1 <- GenerateMilData(positive_dist = 'mvt',
                            negative_dist = 'mvnormal',
                            remainder_dist = 'mvnormal',
                            nbag = 50,
                            nsample = 20,
                            positive_degree = 3,
                            positive_prob = 0.15,
                            positive_mean = rep(0, 5))
df1 <- build_instance_feature(MilData1, seq(0.05, 0.95, length.out = 10))
mdl <- MI_SVM(data = df1, cost = 1, kernel = 'radial')
predictions1_MI <- predict(mdl, newdata = df1, true_bag_info = unique(df1[, 1:2]))
}
\author{
Yifei Liu
}
