% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mildsvm.R
\name{predict.mildsvm}
\alias{predict.mildsvm}
\title{Predict method for 'mildsvm' object}
\usage{
\method{predict}{mildsvm}(object, newX, type = c("prediction", "score"))
}
\arguments{
\item{object}{an object of class 'mildsvm'}

\item{newX}{TODO:}

\item{type}{TODO:}
}
\description{
Predict method for 'mildsvm' object
}
\author{
Sean Kent
}
